(refactor) formatting
Add headers and formatting to all choco owned c# files.

+++ chocolatey/Website/Services/FileSystemFileStorageService.cs
@@ -1,4 +1,22 @@
-﻿using System;
+﻿// Copyright 2011 - Present RealDimensions Software, LLC, the original 
+// authors/contributors from ChocolateyGallery
+// at https://github.com/chocolatey/chocolatey.org,
+// and the authors/contributors of NuGetGallery 
+// at https://github.com/NuGet/NuGetGallery
+//  
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// 
+//   http://www.apache.org/licenses/LICENSE-2.0
+// 
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System;
 using System.Globalization;
 using System.IO;
 using System.Web.Mvc;
@@ -7,8 +25,8 @@ namespace NuGetGallery
 {
     public class FileSystemFileStorageService : IFileStorageService
     {
-        readonly IConfiguration configuration;
-        readonly IFileSystemService fileSystemSvc;
+        private readonly IConfiguration configuration;
+        private readonly IFileSystemService fileSystemSvc;
 
         public FileSystemFileStorageService(
             IConfiguration configuration,
@@ -18,7 +36,7 @@ namespace NuGetGallery
             this.fileSystemSvc = fileSystemSvc;
         }
 
-        static string BuildPath(
+        private static string BuildPath(
             string fileStorageDirectory,
             string folderName,
             string fileName)
@@ -30,30 +48,28 @@ namespace NuGetGallery
             string folderName,
             string fileName)
         {
-            if (String.IsNullOrWhiteSpace(folderName))
-                throw new ArgumentNullException("folderName");
-            if (String.IsNullOrWhiteSpace(fileName))
-                throw new ArgumentNullException("fileName");
+            if (String.IsNullOrWhiteSpace(folderName)) throw new ArgumentNullException("folderName");
+            if (String.IsNullOrWhiteSpace(fileName)) throw new ArgumentNullException("fileName");
 
             var path = BuildPath(configuration.FileStorageDirectory, folderName, fileName);
-            if (!fileSystemSvc.FileExists(path))
-                return new HttpNotFoundResult();
+            if (!fileSystemSvc.FileExists(path)) return new HttpNotFoundResult();
 
             var result = new FilePathResult(path, GetContentType(folderName));
             result.FileDownloadName = new FileInfo(fileName).Name;
             return result;
         }
 
-        static string GetContentType(string folderName)
+        private static string GetContentType(string folderName)
         {
             switch (folderName)
             {
-                case Constants.PackagesFolderName:
+                case Constants.PackagesFolderName :
                     return Constants.PackageContentType;
-                case Constants.DownloadsFolderName:
+                case Constants.DownloadsFolderName :
                     return Constants.OctetStreamContentType;
-                default:
-                    throw new InvalidOperationException(String.Format(CultureInfo.CurrentCulture, "The folder name {0} is not supported.", folderName));
+                default :
+                    throw new InvalidOperationException(
+                        String.Format(CultureInfo.CurrentCulture, "The folder name {0} is not supported.", folderName));
             }
         }
 
@@ -61,24 +77,19 @@ namespace NuGetGallery
             string folderName,
             string fileName)
         {
-            if (String.IsNullOrWhiteSpace(folderName))
-                throw new ArgumentNullException("folderName");
-            if (String.IsNullOrWhiteSpace(fileName))
-                throw new ArgumentNullException("fileName");
+            if (String.IsNullOrWhiteSpace(folderName)) throw new ArgumentNullException("folderName");
+            if (String.IsNullOrWhiteSpace(fileName)) throw new ArgumentNullException("fileName");
 
             var path = BuildPath(configuration.FileStorageDirectory, folderName, fileName);
-            if (fileSystemSvc.FileExists(path))
-                fileSystemSvc.DeleteFile(path);
+            if (fileSystemSvc.FileExists(path)) fileSystemSvc.DeleteFile(path);
         }
 
         public bool FileExists(
             string folderName,
             string fileName)
         {
-            if (String.IsNullOrWhiteSpace(folderName))
-                throw new ArgumentNullException("folderName");
-            if (String.IsNullOrWhiteSpace(fileName))
-                throw new ArgumentNullException("fileName");
+            if (String.IsNullOrWhiteSpace(folderName)) throw new ArgumentNullException("folderName");
+            if (String.IsNullOrWhiteSpace(fileName)) throw new ArgumentNullException("fileName");
 
             var path = BuildPath(configuration.FileStorageDirectory, folderName, fileName);
             return fileSystemSvc.FileExists(path);
@@ -88,16 +99,12 @@ namespace NuGetGallery
             string folderName,
             string fileName)
         {
-            if (String.IsNullOrWhiteSpace(folderName))
-                throw new ArgumentNullException("folderName");
-            if (String.IsNullOrWhiteSpace(fileName))
-                throw new ArgumentNullException("fileName");
+            if (String.IsNullOrWhiteSpace(folderName)) throw new ArgumentNullException("folderName");
+            if (String.IsNullOrWhiteSpace(fileName)) throw new ArgumentNullException("fileName");
 
             var path = BuildPath(configuration.FileStorageDirectory, folderName, fileName);
-            if (fileSystemSvc.FileExists(path))
-                return fileSystemSvc.OpenRead(path);
-            else
-                return null;
+            if (fileSystemSvc.FileExists(path)) return fileSystemSvc.OpenRead(path);
+            else return null;
         }
 
         public void SaveFile(
@@ -105,7 +112,7 @@ namespace NuGetGallery
             string fileName,
             Stream packageFile)
         {
-            if (String.IsNullOrWhiteSpace(folderName))throw new ArgumentNullException("folderName");
+            if (String.IsNullOrWhiteSpace(folderName)) throw new ArgumentNullException("folderName");
             if (String.IsNullOrWhiteSpace(fileName)) throw new ArgumentNullException("fileName");
             if (packageFile == null) throw new ArgumentNullException("packageFile");
 
@@ -115,10 +122,7 @@ namespace NuGetGallery
             if (!fileSystemSvc.DirectoryExists(folderPath)) fileSystemSvc.CreateDirectory(folderPath);
 
             var filePath = BuildPath(configuration.FileStorageDirectory, folderName, fileName);
-            if (fileSystemSvc.FileExists(filePath))
-            {
-                fileSystemSvc.DeleteFile(filePath);
-            }
+            if (fileSystemSvc.FileExists(filePath)) fileSystemSvc.DeleteFile(filePath);
 
             using (var file = fileSystemSvc.OpenWrite(filePath))
             {
